docs: adicionada exemplo e teste unitário aa func delete e docs em formato comentário em funções não exportáveis
